requirements.txt
google-genai
mcp>=1.9.4
fastmcp>=2.8.1